# Avoid unnecessarily disabling CUDA graphs