@slaren 2024-05-15
| GPU         | Model         | Test        |   t/s master |   t/s ag_fix_consecutive_update_counter_increment |   Speedup |
|:------------|:--------------|:------------|-------------:|--------------------------------------------------:|----------:|
| RTX 3090 Ti | llama 7B Q4_0 | pp512       |      4938.82 |                                           4940.78 |      1.00 |
| RTX 3090 Ti | llama 7B Q4_0 | tg128       |       164.12 |                                            163.80 |      1.00 |
| RTX 3090 Ti | llama 7B Q4_0 | pp512+tg128 |       599.39 |                                            691.16 |      1.15 |