@slaren 2024-05-15
The benchmark is also failing on master so I don't think it is caused by this.