@skoulik 2024-05-29
@agray3 

Any chance of enabling graphs on Turing architecture?
I've tried it on RTX 2080 Ti (arch 7.5) and the graphs seem to be supported.