@agray3 2024-05-30
> @agray3
> 
> Any chance of enabling graphs on Turing architecture? I've tried it on RTX 2080 Ti (CC 7.5) and the graphs seem to be supported.

@skoulik The reason that CUDA graphs are disabled for older GPUs is related to performance - have you tested performance with and without CUDA graphs on your Turing GPU? From my test (on Quadro RTX 8000), it is actually a bit slower with graphs. To remove the restriction, just comment out the line https://github.com/agray3/llama.cpp/blob/d5c05821f3c3d6cabe8ac45776fe0ecb0da13eca/ggml-cuda.cu#L2533 

If there is a benefit in your case, we could maybe add  GGML_CUDA_ENABLE_GRAPHS environment variable to override the default behaviour (noting we currently have the opposite GGML_CUDA_DISABLE_GRAPHS).