@skoulik 2024-05-30
> > @agray3
> 
> @skoulik The reason that CUDA graphs are disabled for older GPUs is related to performance - have you tested performance with and without CUDA graphs on your Turing GPU? 

Actually not, because I seem cannot be able to. What I found in my usecase is that the graps keep being disabled because of cuda_ctx->cuda_graph->disable_due_to_too_many_updates a few lines lower. This is when I run server inference with llama 3 model.  Can you tell me the usecase(s) where it is supposed to work? I'll be happy to test.

>From my test (on Quadro RTX 8000), it is actually a bit slower with graphs.

Understood.

> To remove the restriction, just comment out the line https://github.com/agray3/llama.cpp/blob/d5c05821f3c3d6cabe8ac45776fe0ecb0da13eca/ggml-cuda.cu#L2533
> 
That's what I did.