@skoulik 2024-05-31
> > Actually not, because I seem cannot be able to. What I found in my usecase is that the graps keep being disabled because of cuda_ctx->cuda_graph->disable_due_to_too_many_updates a few lines lower. This is when I run server inference with llama 3 model. Can you tell me the usecase(s) where it is supposed to work? I'll be happy to test.
> 
> I'm not sure why that is happening - can you please tell me your exact command to allow me to reproduce?

Of course, here is the command I run:
```
main.exe -m models\Meta-Llama-3-8B-Instruct_64K_Q8_0.gguf --ctx-size 65536 --n-gpu-layers 33 -t 1 --flash-attn --override-kv tokenizer.ggml.pre=str:llama3 -s 1
```

The quants downloaded from https://huggingface.co/NurtureAI/Meta-Llama-3-8B-Instruct-64k-GGUF

I've added logging and confirm that all calls end up in disable_due_to_too_many_updates. Not a single graph launch, unfortunately.