@agray3 2024-05-31
@skoulik I tried with your exact command but unfortunately I cannot reproduce - it uses CUDA graphs for me as expected (including on Turing if I remove the restriction). All I can suggest is that you add some print statements to the code to trace back exactly what is causing the `cuda_graph_update_required` flag to be set true across multiple consecutive tokens, in your case.