@skoulik 2024-05-31
@agray3 ,
I've done some more debugging.
On startup I got a bunch of 
```
ggml_backend_cuda_graph_compute: disabling CUDA graphs due to batch size > 1 [ffn_inp-0] [4096 2 1 1]
ggml_backend_cuda_graph_compute: disabling CUDA graphs due to batch size > 1 [ffn_inp-1] [4096 2 1 1]
...
ggml_backend_cuda_graph_compute: disabling CUDA graphs due to batch size > 1 [ffn_inp-30] [4096 2 1 1]
```
I suppose these are expected.
Then it gets more interesting:
Very first pair of calls to cudaGraphExecUpdate() and cudaGraphLaunch() are fine.
Next call to cudaGraphExecUpdate() fails:
```
stat  cudaErrorGraphExecUpdateFailure (910)   cudaError
result_info:
result  cudaGraphExecUpdateErrorTopologyChanged (2)
errorNode       0x0000000000000000 <NULL>
errorFromNode   0x0000000000000000 <NULL>
```
Then, after 4 retries because of number_consecutive_updates>=4 it sets disable_due_to_too_many_updates = true.

Now, I suppose the question is, what's wrong with graph topology? Any idea on how to debug?