@skoulik 2024-05-31
@agray3 ,
On a second thought, the error cannot be "expected", because it prevents the just captured graph from executing. I reckon the patch that I proposed is valid and should be integrated. What do you think?