@skoulik 2024-05-31
As for the root causes of the cuda_graph_update_required.
The first batch of updates mentioned above is caused by:
```
if (node->op == GGML_OP_ADD && node->src[1] && node->src[1]->ne[1] > 1) 
```
node->src[1]->ne[1] == 2. Which is OK, I believe.

Now, when the generations starts, the updates are caused by: has_matching_properties == false, by thus check: node->src[i]->data != graph_node_properties->src_address[i]
```
 	main.exe!ggml_graph_node_has_matching_properties(ggml_tensor * node=0x0000021415569da0, ggml_graph_node_properties * graph_node_properties=0x000002104502de00) Line 2506	C++
>	main.exe!ggml_backend_cuda_graph_compute(ggml_backend * backend=0x00000210476b5220, ggml_cgraph * cgraph=0x00000210467da3c0) Line 2566	C++
 	main.exe!ggml_backend_graph_compute_async(ggml_backend * backend=0x00000210476b5220, ggml_cgraph * cgraph=0x00000210467da3c0) Line 282	C
 	main.exe!ggml_backend_sched_compute_splits(ggml_backend_sched * sched=0x0000021415868070) Line 1667	C
```
```
    for (int i = 0; i < GGML_MAX_SRC; i++) {
        if (node->src[i] &&
            node->src[i]->data != graph_node_properties->src_address[i] &&
            node->op != GGML_OP_CPY &&
            node->op != GGML_OP_VIEW
        ) {
            return false;
        }
```
```
		node->src[i]->data	0x0000001916f60000	void *
```
```
		graph_node_properties->src_address[i]	0x0000001909258000	void *
```
Not sure what this means and why it works for you,