@agray3 2024-05-31
> I reckon the patch that I proposed is valid and should be integrated. What do you think?

This change always destroys any instantiated graph, but the instantiated graph is required for the graph update mechanism, which involves updating a previously instantiated graph with a newly captured graph to create an updated instance.

>Not sure what this means and why it works for you,

Yeah it's weird we are seeing different behaviour. Can you please share the full standard output you get from 
```
main.exe -m models\Meta-Llama-3-8B-Instruct_64K_Q8_0.gguf --ctx-size 65536 --n-gpu-layers 33 -t 1 --flash-attn --override-kv tokenizer.ggml.pre=str:llama3 -s 1
```
and I'll cross check with what I get,?